+++ HYDownloadManagerPod/IADownloadManager.m
@@ -15,7 +15,7 @@
 @property (nonatomic, strong) NSMutableDictionary *downloadOperations;
 @property (nonatomic, strong) NSMutableDictionary *downloadHandlers;
 
-- (void)removeHandlerWithTag:(int)tag;
+- (void)removeHandlerWithTag:(NSInteger)tag;
 
 @end
 
@@ -112,7 +112,7 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
 
 - (void) attachNewHandlerWithProgressBlock:(IAProgressBlock)progressBlock
                            completionBlock:(IACompletionBlock)completionBlock
-                                       tag:(int)tag
+                                       tag:(NSInteger)tag
                                      toURL:(NSURL*)url
 {
     //unlink the tag from the urls
@@ -177,12 +177,12 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
 #pragma mark IADownloadHandlers and IADownloadOperation Helpers
 #pragma mark -
 
-- (void)removeHandlerWithURL:(NSURL*)url tag:(int)tag
+- (void)removeHandlerWithURL:(NSURL*)url tag:(NSInteger)tag
 {
     NSMutableArray *handlers = [self.downloadHandlers objectForKey:url];
     if (handlers)
     {
-        for (int i = handlers.count - 1; i >= 0; i-- )
+        for (NSInteger i = handlers.count - 1; i >= 0; i-- )
         {
             IADownloadHandler *handler = handlers[i];
             
@@ -199,7 +199,7 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
     NSMutableArray *handlers = [self.downloadHandlers objectForKey:url];
     if (handlers)
     {
-        for (int i = handlers.count - 1; i >= 0; i-- )
+        for (NSInteger i = handlers.count - 1; i >= 0; i-- )
         {
             IADownloadHandler *handler = handlers[i];
             
@@ -211,14 +211,14 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
     }
 }
 
-- (void)removeHandlerWithTag:(int)tag
+- (void)removeHandlerWithTag:(NSInteger)tag
 {
-    for (int i = self.downloadHandlers.allKeys.count - 1; i >= 0; i-- )
+    for (NSInteger i = self.downloadHandlers.allKeys.count - 1; i >= 0; i-- )
     {
         id key = self.downloadHandlers.allKeys[i];
         NSMutableArray *array = [self.downloadHandlers objectForKey:key];
         
-        for (int j = array.count - 1; j >= 0; j-- )
+        for (NSInteger j = array.count - 1; j >= 0; j-- )
         {
             IADownloadHandler *handler = array[j];
             if (handler.tag == tag)
@@ -231,12 +231,12 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
 
 - (void)removeHandlerWithListener:(id)listener
 {
-    for (int i = self.downloadHandlers.allKeys.count - 1; i >= 0; i-- )
+    for (NSInteger i = self.downloadHandlers.allKeys.count - 1; i >= 0; i-- )
     {
         id key = self.downloadHandlers.allKeys[i];
         NSMutableArray *array = [self.downloadHandlers objectForKey:key];
         
-        for (int j = array.count - 1; j >= 0; j-- )
+        for (NSInteger j = array.count - 1; j >= 0; j-- )
         {
             IADownloadHandler *handler = array[j];
             if (handler.delegate == listener)
@@ -296,7 +296,7 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
 }
 
 + (void) stopDownloadingItemWithURL:(NSURL*)url
-                             andTag:(int)tag
+                             andTag:(NSInteger)tag
 {
     [self.instance removeHandlerWithURL:url tag:tag];
     
@@ -311,7 +311,7 @@ void (^globalCompletionBlock)(BOOL success, id response, NSURL *url, IADownloadM
     [[[self instance].downloadOperations objectForKey:url] stop];
 }
 
-+ (int)listenerCountForUrl:(NSURL *)url
++ (NSInteger)listenerCountForUrl:(NSURL *)url
 {
     NSMutableArray *handlers = [[self instance].downloadHandlers objectForKey:url];
     return handlers.count;
